requirements.txt
fastapi
python-multipart
uvicorn
openai
unstructured
supabase
python-jose[cryptography]
passlib[bcrypt]
python-dotenv

